+++ CHANGELOG.md
@@ -11,8 +11,10 @@ Visit the package at [pub.dev](https://pub.dev/packages/solidpod).
 
 ## 1.0
 
-+ Switch public to private sharing [1.0.15 20260728 jesscmoore]
-+ Preserve encryption state on overwrite [1.0.14 20260730 jesscmoore]
++ Switch public to private sharing [1.0.17 20260811 jesscmoore]
++ Preserve encryption state on overwrite [1.0.16 20260811 jesscmoore]
++ Fix deleting a large file stored as a single chunk [1.0.15 20260809 gjw]
++ Remove debug output when reading a large file [1.0.14 20260809 gjw]
 + Add load test to the example app [1.0.13 20260702 tonypioneer]
 + Migrate TEMPLATE to solidui [1.0.12 20260629 tonypioneer]
 + Support profile editing [1.0.11 20260626 tonypioneer]